rename script and add fix startrack
