+++ scripts/tests/test_fix_trk.py
@@ -3,6 +3,5 @@
 
 
 def test_help_option(script_runner):
-    ret = script_runner.run('scil_fix_dsi_studio_trk.py', '--help')
+    ret = script_runner.run('scil_fix_trk.py', '--help')
     assert ret.success
-
